clean up architecture, tests
