+++ src/pynxtools/annotator/README.md
@@ -1,0 +1,8 @@
+# NeXus file reader and debugger
+
+This utility outputs a debug log for a given NeXus file by annotating the data and metadata entries with the schema definitions from the respective NeXus base classes and application definitions to which the file refers to.
+
+Here you can find more information about the tool:
+
+- [Usage guide](https://fairmat-nfdi.github.io/pynxtools/learn/pynxtools/nexus-validation.html#pynx-read-nexus-file-reader-and-debugger)
+- [API documentation](https://fairmat-nfdi.github.io/pynxtools/reference/cli-api.html#pynx-read)
