+++ src/pynxtools/nexus/README.md
@@ -1,8 +1,1 @@
-# NeXus file reader and debugger
-
-This utility outputs a debug log for a given NeXus file by annotating the data and metadata entries with the schema definitions from the respective NeXus base classes and application definitions to which the file refers to.
-
-Here you can find more information about the tool:
-
-- [Usage guide](https://fairmat-nfdi.github.io/pynxtools/learn/pynxtools/nexus-validation.html#read_nexus-nexus-file-reader-and-debugger)
-- [API documentation](https://fairmat-nfdi.github.io/pynxtools/reference/cli-api.html#read-nexus)
+# NeXus utilities
