+++ tests/annotator/README.md
@@ -1,0 +1,4 @@
+# NeXus annotator tests
+
+This is the place for storing code for tests which probe key functionalities of the ``annotator`` module.
+The ``annotator`` module outputs a debug log for a given NeXus file by annotating the data and metadata entries with the schema definitions from the respective NeXus base classes and application definitions to which the file refers to.
