+++ tests/nexus/README.md
@@ -1,9 +1,5 @@
-This is the place for storing code for tests which probe key functionalities of the nexus(checker).
-Nexuschecker is a core package of which implements functionalities for traversing the XML tree
-that is behind the hierarchical data model of NeXus. The nexuschecker `nexus.py` can also be used
+# NeXus utilities
+
+This is the place for storing code for tests which probe key functionalities of the nexus module.
+``nexus`` is a core module which implements functionalities for traversing the XML tree that is behind the hierarchical data model of NeXus. The nexuschecker `nexus.py` can also be used
 as a standalone tool to check how an instance of a NeXus/HDF5 file compares to a given application definition.
-
-<span style="color:red">Sandor, please reactivate here your checker code from the master branch and maybe add a small description here.</span>
-
-## Contact person in FAIRmat for these tests
-Sándor Brockhauser
